Fast pointer patching and txt_unpack for "loaded" data
For my benchmark test (8,5MB of binary data with a lot of pointers in unions), it used to take 1.92 seconds to patch the pointers on my machine. After this change it takes 0.00056 seconds to do the same thing.
I'm changing the base address which pointer offsets are calculated from to be the header address instead of the instance root. This means that DL_INSTANCE_VERSION needed to be bumped to 2. There is backwards compatibility added so old data can be read and resaved.
I'm also adding more CMake rules, to simplify building and testing on other platforms than Windows.

+++ src/dl_binary_writer.h
@@ -174,7 +174,7 @@ static inline void dl_binary_writer_write_array( dl_binary_writer* writer, const
 }
 
 // val is expected to be in host-endian!!!
-static inline void dl_binary_writer_write_ptr( dl_binary_writer* writer, size_t val )
+static inline void dl_binary_writer_write_ptr( dl_binary_writer* writer, uint64_t val )
 {
 	if( writer->target_endian != DL_ENDIAN_HOST )
 	{
